requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
requests==2.32.3
lxml==5.3.0
pydantic==2.9.2
loguru==0.7.2